looks super awesome but we're hitting an 'out of resources' error at some magic iteration: at res: 640x480x1 & num: 640x480, at ~ frame 275, AND at res: 640x480x1 & num: 640x480/4, at ~ frame 500 we see a resource failure. I've ruled out OOM (it seems I'm barely using 500 mb VRAM -- out of 4 gb!) pyopencl.RuntimeError: clWaitForEvents failed: out of resources we could be hitting a weird nvidia api limit -- voodoo!! -- but I rather suspect this is a segfault, due to: 1) running the same code w/o the physics engine we don't see the same limit (warning: halting problem ahead!!). 2) the consistency but imprecision of the error in runtime suggests a probability of its encounter. with point '2' in mind: I've included a catch for pyopencl.RuntimeError that will print out the whooooole numpy array (not very useful) and I've included some logic for persisting the array + the resolution to an hd5 file. the hd5 code is totally untested - but the goal is to be able to save and restore the world at a given runtime, so I can try tweaking the kernal code until it passes.
NOTE: this belongs somewhere above, but w/e:
When I initially implemented the try:except for the pyopencl.RuntimeError, I was simply swallowing the error and re-entering the loop-scope. Because we excepted, this meant we *were* retrying the same population time and again - with the same RESULTS!
this suggests that what I'm encountering is most likely a stack overflow due to some inscrutable (read: bad) code.
Hopefully, the hd5 persistence code will give me the ability to test this behavior in detail.

Further NOTE:
if someone could invent a good OpenCL debugger for Nvidia...man...the world'd be your oyster